+++ HPXMLtoOpenStudio/measure.rb
@@ -137,6 +137,16 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
         fail 'Weather station EPW filepath has different values across dwelling units.'
       end
 
+      if building_id == 'ALL'
+        if hpxml.buildings.map { |hpxml_bldg| hpxml_bldg.batteries.size }.sum > 0
+          # FIXME: Figure out how to allow this. If we allow it, update docs and hpxml_translator_test.rb too.
+          # Batteries use "TrackFacilityElectricDemandStoreExcessOnSite"; to support modeling of batteries in whole
+          # SFA/MF building simulations, we'd need to create custom meters with electricity usage *for each unit*
+          # and switch to "TrackMeterDemandStoreExcessOnSite".
+          fail 'Modeling batteries for whole SFA/MF buildings is not currently supported.'
+        end
+      end
+
       # Apply HPXML defaults upfront; process schedules & emissions
       hpxml_sch_map = {}
       check_emissions_references(hpxml.header, hpxml_path)